test/requirements.txt
paho-mqtt
colorlog
pretenders
